+++ languagetool-standalone/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## 6.0-SNAPSHOT (release planned for 2022-12-29)
 
+#### Catalan
+  * added and improved rules
+  * updated dictionary (catalan-pos-dict-2.21)
+
 #### English
   * added and improved rules
   * updated en_GB spellchecker dictionary from https://github.com/marcoagpinto/aoo-mozilla-en-dict (Version 2022.12.01, v.3.1.5)